Feat: Make create order API, complete hibernate and mapping entity

+++ frontend/frontend/src/main/java/com/example/SelfOrderingRestaurant/Repository/DishRepository.java
@@ -3,5 +3,5 @@ package com.example.SelfOrderingRestaurant.Repository;
 import com.example.SelfOrderingRestaurant.Entity.Dish;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DishRepository extends JpaRepository<Dish, Long> {
+public interface DishRepository extends JpaRepository<Dish, Integer> {
 }
