+++ frontend/frontend/src/main/java/com/example/SelfOrderingRestaurant/Repository/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.example.SelfOrderingRestaurant.Repository;
 import com.example.SelfOrderingRestaurant.Entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
 }
